+++ components/Common/index.ts
@@ -1,0 +1,5 @@
+import Header from './Header/Header';
+import Modal from './Modal/Modal';
+import Navigation from './Navigation/Navigation';
+
+export { Header, Modal, Navigation };
