+++ pages/_app.tsx
@@ -1,8 +1,8 @@
 import { RecoilRoot, RecoilEnv } from 'recoil';
 import type { AppProps } from 'next/app';
-import Navigation from '@/components/Common/Navigation/Navigation';
 import '@/styles/normalize.scss';
 import '@/styles/fonts.scss';
+import { Navigation } from '@/components/Common';
 
 export default function App({ Component, pageProps }: AppProps) {
   RecoilEnv.RECOIL_DUPLICATE_ATOM_KEY_CHECKING_ENABLED = false;
